requirements/common.txt
Django
pillow
git+https://github.com/jhs512/django-pydenticon
django_bootstrap5
mysqlclient
django-dotenv
django_bootstrap5
lazy-string
requests
djangorestframework
drf-writable-nested
django-cors-headers
django-debug-toolbar
djangorestframework-simplejwt